Split `FunctionBindgen` into its own file
